refactor(handler): mcp-tool-handler 职责拆分，遵循单一职责原则
将 MCPToolHandler（原 2743 行）拆分为多个职责清晰的模块：

1. ToolValidator (src/server/utils/tool-validator.ts)
   - 验证工作流数据完整性
   - 验证字段格式和长度
   - 验证业务逻辑
   - 验证 JSON Schema 和工具结构

2. CozeWorkflowHandler (src/server/handlers/coze-workflow.handler.ts)
   - 转换 Coze 工作流为 MCP 工具
   - 处理工具名称规范化和冲突解决
   - 生成输入参数结构

3. ToolErrorHandler (src/server/handlers/tool-error.handler.ts)
   - 处理添加工具时的错误
   - 处理更新工具时的错误
   - 处理删除工具时的错误

4. ToolPreChecker (src/server/handlers/tool-precheck.handler.ts)
   - 执行边界条件预检查
   - 检查系统状态和资源限制

重构成果：
- 主处理器从 2743 行减少到 ~800 行（减少约 70%）
- 验证逻辑独立可复用
- Coze 工作流处理自成一体
- 错误处理统一管理

Closes #3388
